make events optional

+++ lib/src/tree.ts
@@ -10,6 +10,7 @@ import { Config, PillarConfig, Options } from './config';
 import { Events } from './events';
 import sort from 'fast-sort'
 import { errors } from './errors';
+import Delta from "quill-delta"
 import "../assets/tree.css"
 
 const firstCardText = ""
@@ -98,7 +99,7 @@ export class Tree implements RedomComponent {
             let firstNode: Node = {
                 uid: 0,
                 pos: new Pos(),
-                text: firstCardText,
+                content: firstCardText,
             }
             this.appendNode(firstNode)
             this.cardIndexer.push(firstNode.pos)
@@ -135,13 +136,7 @@ export class Tree implements RedomComponent {
         if (options && options.events) {
             this.event = options.events
         } else {
-            this.event = {
-                onNewNode: (uid: number, pos: Pos) => { },
-                onMoveNode: (oldPos: Pos, newPos: Pos) => { },
-                onModifyNode: (node: Node) => { },
-                onDeleteNode: (node: Node) => { },
-                onSelectNode: (node: Node) => { },
-            }
+            this.event = {}
         }
 
         // begin by focusing on the first card
@@ -242,11 +237,11 @@ export class Tree implements RedomComponent {
     // has an invalid position. Moves cards in the same family that are
     // below by one index down. Will not cause the matching insertNode
     // event to fire
-    insertNode(pos: Pos, focus: boolean = false, text: string = ""): Node {
+    insertNode(pos: Pos, focus: boolean = false, content: string | Delta = ""): Node {
         let node = {
             uid: this.cardIndexer.length,
             pos: pos,
-            text: text
+            content: content
         }
 
         this.validateNewNode(node) 
@@ -295,7 +290,7 @@ export class Tree implements RedomComponent {
         let toNode = {
             uid: fromNode.uid,
             pos: to.copy(),
-            text: fromNode.text,
+            content: fromNode.content,
         }
 
         // validate the the position where the node shall be moved to (we treat it as if adding a new node)
@@ -443,7 +438,7 @@ export class Tree implements RedomComponent {
                     }
 
                     if (withText) {
-                        result += "\t\t\t" + card.node().text
+                        result += "\t\t\t" + card.editor.getText()
                     }
                 })
             })
@@ -598,13 +593,15 @@ export class Tree implements RedomComponent {
                     if (this.card.atStart()) {
                         if (this.shiftMode) {
                             this.createAbove()
-                            this.event.onNewNode(this.card.id(), this.current.copy())
+                            if (this.event.onNewNode)
+                                this.event.onNewNode(this.card.id(), this.current.copy())
                         } else if (this.altMode) {
                             this.shiftUpwards()
                         } else {
                             this.up()
                             this.card.focus()
-                            this.event.onSelectNode(this.card.node())
+                            if (this.event.onSelectNode)
+                                this.event.onSelectNode(this.card.node())
                         }
                     }
                     break;
@@ -612,13 +609,15 @@ export class Tree implements RedomComponent {
                     if (this.card.atEnd()) {
                         if (this.shiftMode) {
                             this.createBelow()
-                            this.event.onNewNode(this.card.id(), this.current.copy())
+                            if (this.event.onNewNode)
+                                this.event.onNewNode(this.card.id(), this.current.copy())
                         } else if (this.altMode) {
                             this.shiftDownwards()
                         } else {
                             this.down()
                             this.card.focusStart()
-                            this.event.onSelectNode(this.card.node())
+                            if (this.event.onSelectNode) 
+                                this.event.onSelectNode(this.card.node())
                         }
                     }
                     break;
@@ -626,11 +625,13 @@ export class Tree implements RedomComponent {
                     if (this.card.atEnd()) {
                         if (this.shiftMode) {
                             this.createChild()
-                            this.event.onNewNode(this.card.id(), this.current.copy())
+                            if (this.event.onNewNode)
+                                this.event.onNewNode(this.card.id(), this.current.copy())
                         } else {
                             this.down()
                             this.card.focusStart()
-                            this.event.onSelectNode(this.card.node())
+                            if (this.event.onSelectNode)
+                                this.event.onSelectNode(this.card.node())
                         }
                     }
                     break;
@@ -638,14 +639,16 @@ export class Tree implements RedomComponent {
                     if (this.card.atStart()) {
                         this.left()
                         this.card.focusStart()
-                        this.event.onSelectNode(this.card.node())
+                        if (this.event.onSelectNode)
+                            this.event.onSelectNode(this.card.node())
                     }
                     break;
                 case "Backspace":
                     if (this.card.backspace()) {
                         let node = this.getCard().node()
                         this.deleteNode(this.current)
-                        this.event.onDeleteNode(node)
+                        if (this.event.onDeleteNode)
+                            this.event.onDeleteNode(node)
                     }
                     break;
                 case "Escape":
@@ -658,41 +661,49 @@ export class Tree implements RedomComponent {
             case "ArrowUp":
                 if (this.shiftMode) {
                     this.createAbove()
-                    this.event.onNewNode(this.card.id(), this.current.copy())
+                    if (this.event.onNewNode)
+                        this.event.onNewNode(this.card.id(), this.current.copy())
                 } else if (this.altMode) {
                     this.shiftUpwards()
                 } else {
                     this.up();
-                    this.event.onSelectNode(this.card.node())
+                    if (this.event.onSelectNode)
+                        this.event.onSelectNode(this.card.node())
                 }
                 break;
             case "ArrowDown":
                 if (this.shiftMode) {
                     this.createBelow()
-                    this.event.onNewNode(this.card.id(), this.current.copy())
+                    if (this.event.onNewNode)
+                        this.event.onNewNode(this.card.id(), this.current.copy())
                 } else if (this.altMode) {
                     this.shiftDownwards()
                 } else {
                     this.down();
-                    this.event.onSelectNode(this.card.node())
+                    if (this.event.onSelectNode)
+                        this.event.onSelectNode(this.card.node())
                 }
                 break;
             case "ArrowLeft":
                 if (this.shiftMode) {
                     this.createParent()
-                    this.event.onNewNode(this.card.id(), this.current.copy())
+                    if (this.event.onNewNode)
+                        this.event.onNewNode(this.card.id(), this.current.copy())
                 } else {
                     this.left();
-                    this.event.onSelectNode(this.card.node())
+                    if (this.event.onSelectNode)
+                        this.event.onSelectNode(this.card.node())
                 }
                 break;
             case "ArrowRight":
                 if (this.shiftMode) {
                     this.createChild()
-                    this.event.onNewNode(this.card.id(), this.current.copy())
+                    if (this.event.onNewNode)
+                        this.event.onNewNode(this.card.id(), this.current.copy())
                 } else {
                     this.right();
-                    this.event.onSelectNode(this.card.node())
+                    if (this.event.onSelectNode)
+                        this.event.onSelectNode(this.card.node())
                 }
                 break;
             case "Enter":
@@ -701,7 +712,8 @@ export class Tree implements RedomComponent {
             case "Backspace":
                 let node = this.getCard().node()
                 this.deleteNode(this.current)
-                this.event.onDeleteNode(node)
+                if (this.event.onDeleteNode)
+                    this.event.onDeleteNode(node)
                 break;
         }
 
@@ -770,7 +782,8 @@ export class Tree implements RedomComponent {
         let newPos = pos.above(this.pillars[pos.depth])
         if (newPos.isNotNull()) {
             this.moveNode(pos, newPos, true)
-            this.event.onMoveNode(pos, newPos)
+            if (this.event.onMoveNode)
+                this.event.onMoveNode(this.card.id(), pos, newPos)
         }
     }
 
@@ -780,7 +793,8 @@ export class Tree implements RedomComponent {
         let newPos = pos.below(this.pillars[pos.depth])
         if (newPos) {
             this.moveNode(pos, newPos, true)
-            this.event.onMoveNode(pos, newPos)
+            if (this.event.onMoveNode)
+                this.event.onMoveNode(this.card.id(), pos, newPos)
         }
     }
